Updated /text-to-speech to return test.mp3
